requirements.txt
python-dateutil>=2.8.1
numpy>=1.18.5
pytz>=2020.1
six>=1.5

pandas==1.4.2
ffmpeg-python==0.2.0